update header component
