+++ src/app/features/public/pages/about/page/about.component.html
@@ -1,4 +1,5 @@
 <app-loader></app-loader>
+<app-header></app-header>
 <app-titulo [title]="title"></app-titulo>
 <div class="p-4 p-md-5 mb-4 rounded border shadow m-md-5">
   <div class="col-md-8 px-0">
