+++ src/app/features/public/pages/activities-view/activities-view.component.html
@@ -1,3 +1,4 @@
+<app-header></app-header>
 <section class="container">
   <app-titulo [title]="'Actividades'"></app-titulo>
   <div class="row row-cols-lg-3 row-cols-md-2 g-3 flex-row">
