+++ src/app/features/public/pages/contact-form/contact-form.component.html
@@ -98,7 +98,7 @@
           Enviar consulta
         </button>
       </form>
-      <a [routerLink]=""><button
+      <a [routerLink]="['']"><button
         class="btn border border-dark p-3 mt-3 radius"
         type="submit"
       >
