--- src/app/features/public/pages/header/header.component.html
@@ -1,34 +1,0 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light">
-  <div class="container-fluid font">
-    <a class="navbar-brand" href="#"
-      ><img
-        src="../../../../../assets/images/LOGO-SOMOS MAS.png"
-        alt="Logo somos mas"
-    /></a>
-    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-      <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-        <li class="nav-item">
-          <a class="nav-link" aria-current="page" routerLink="/" routerLinkActive="activeButton">Inicio</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" routerLink="/nosotros" routerLinkActive="activeButton">Nosotros</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" routerLink="/contact-form" routerLinkActive="activeButton">Contacto</a>
-        </li>
-        <li class="nav-item dropdown">
-          <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-            Campañas
-          </a>
-          <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-            <li><a class="dropdown-item" href="#">Recolección de materiales escolares</a></li>
-            <li><a class="dropdown-item" href="#">Recolección de Juguetes</a></li>
-          </ul>
-        </li>
-      </ul>
-    </div>
-  </div>
-</nav>
