+++ src/app/features/public/pages/home/page/home.component.html
@@ -1,4 +1,5 @@
 <app-loader></app-loader>
+<app-header></app-header>
 <div class="container d-flex justify-content-center aling-items-center flex-column">
     <app-hero [organizationData]="organizationData"></app-hero>
     <div class="w-100 d-flex overflow-hidden justify-content-center aling-items-center flex-column">
