+++ src/app/features/public/public.module.ts
@@ -14,7 +14,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { PublicRoutingModule } from './public-routing.module';
 import { AngularFireModule } from '@angular/fire/compat';
 import { environment } from 'src/environments/environment.prod';
-import { HeaderComponent } from './pages/header/header.component';
+
 
 
 @NgModule({
@@ -24,13 +24,11 @@ import { HeaderComponent } from './pages/header/header.component';
     AuthFormComponent,
     ContactFormComponent,
     ActivitiesViewComponent,
-    HeaderComponent,
   ],
   exports: [
     ContactFormComponent,
     LoginFormComponent,
     RegisterFormComponent,
-    HeaderComponent
   ],
   imports: [
     CommonModule,
