+++ src/app/shared/header/header.component.html
@@ -1,0 +1,19 @@
+<nav class="navbar navbar-expand-lg navbar-light bg-light">
+  <div class="container-fluid font">
+    <a class="navbar-brand" href="#"
+      ><img
+        src="../../../../../assets/images/LOGO-SOMOS MAS.png"
+        alt="Logo somos mas"
+    /></a>
+    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+      <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+        <li class="nav-item" *ngFor="let menu of menuList">
+          <a class="nav-link" [routerLink]="menu.link" routerLinkActive="activeButton">{{menu.name}}</a>
+        </li>
+      </ul>
+    </div>
+  </div>
+</nav>
