+++ src/app/shared/shared.module.ts
@@ -14,6 +14,8 @@ import { LazyLoadComponent } from './components/layouts/lazy-load/lazy-load.comp
 import { SkeletonComponent } from "./components/layouts/skeleton/skeleton/skeleton.component";
 import { NgxSkeletonLoaderModule } from "ngx-skeleton-loader";
 import { NotFoundComponent } from './components/not-found/not-found.component';
+import { HeaderComponent } from "./header/header.component";
+
 
 
 @NgModule({
@@ -28,6 +30,7 @@ import { NotFoundComponent } from './components/not-found/not-found.component';
     ProgressBarComponent,
     LoaderComponent,
     NotFoundComponent,
+    HeaderComponent,
   ],
   imports: [
     CommonModule,
@@ -46,6 +49,7 @@ import { NotFoundComponent } from './components/not-found/not-found.component';
     SkeletonComponent,
     ProgressBarComponent,
     LoaderComponent,
+    HeaderComponent,
   ]
 })
 export class SharedModule { }
